x11-misc/lightdm: Remove QT5 support
While upstream is working on QT6 support being added to lightdm it is decided to drop QT5 in the meantime as we have no supporting QT5 greeter in ::gentoo and was unable to find one in any third party repos.

Once https://github.com/canonical/lightdm/pull/373 is merged support will be added back.

Closes: https://bugs.gentoo.org/953261
Signed-off-by: Ian Jordan <immoloism@gmail.com>