profiles/arch/arm64: package.use.mask still unsupported wine USE
Not adding wine-proton given really no intention to keyword that one
on arm64, rather unlikely Valve tests it there and they will not offer
support if something is broken.

Note USE=mono is use.mask'ed on arm64 either way, but explicitly list
it given it's not about the same packages for wine. In the future may
instead need to unmask it here to allow a future ported wine-mono.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>