dev-python/pysvn: Make it build again
A recent update of distutils-r1 (commit 67f008f6acc0?) broke this:
 * ERROR: dev-python/pysvn-1.9.22-r1::gentoo failed (compile phase):
 *   No wheel installed in python_compile(), did you call distutils-r1_python_compile?

Generally, this package doesn't seem to play well with that eclass,
and a more thorough fix by somebody with better knowledge of Python's
build intricacies (and the convoluted eclass code) may be needed.

In any case, it fixes building of the package for now, and its only
reverse dependency dev-vcs/svneverever appears to work fine.

Closes: https://bugs.gentoo.org/953830
Signed-off-by: Ulrich Müller <ulm@gentoo.org>