net-p2p/qbittorrent: add 5.0.5 with improvements
- remove USE=systemd since it was just for installing files
- sort RDEPEND
- bump libtorrent-rasterbar version due to the drop of :1 (see below)
- make MULTIBUILD_VARIANTS assignment declarative & move it
  to src_configure
- sort mycmakeargs & change comment style
- improve multibuild logic comment

qbittorrent supports both 1.x and 2.x series of libtorrent-rasterbar.
We couldn't reflect that by using `||` as it doesn't support `:=`.
The solution was to define a minimum 1.x version so it was at least
halfway right. 1.x series was dropped from ::gentoo so I have
changed the minimum version to 2.x, making it fully correct.

Signed-off-by: Filip Kobierski <fkobi@pm.me>