net-print/cups: remove pointless dependency on sys-apps/attr
This dates back to a completely inscrutable commit
41e289137d3b5d3e7460458cbe77d8b4000b594a in gentoo-historical, dating
back to 2007, in which USE=acl was initially added.

> I wrote a little patch to fix the automagic dependency on sys-apps/acl
> and sys-apps/attr

cups does, genuinely, have an --enable-acl, which is, since that time,
genuinely hooked up via USE and passed to econf. And... it appropriately
depends on sys-apps/acl to implement it. Why did it also depend on attr
as well?  Who knows. acl depends on attr internally (thus the dependency
is already satisfied, and no slot dep) but doesn't and didn't expose it.
Nor did cups attempt to directly link to attr.

Maybe it was some kind of cargo-culted belief that acl and attr simply
"always go together"?

No revbump, this package won't be depcleaned anyway so forcing a rebuild
is nearly as pointless has the 20-year-old dependency itself.

Bug: https://bugs.gentoo.org/192294
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>