dev-qt/qtbase: allow auto-detecting -mno-direct-extern-access
(ask Sam if want the reasoning, haven't read much on this topic)

Not really tested beside checking that the option is passed.

Note that this won't be "fully" used unless also rebuild all of
dev-qt/* after upgrade to qtbase-6.9.0-r1 given the option is
passed by the qtbase defautls to each Qt modules.

Part of why not adding it to 6.8.3 as it'll be difficult to
guarantee what the users actually have without bunch of revbumps
(not so bad for 6.9.0 given mask), will propagate properly when
add 6.9.1 unmasked.

Unsure if a better way than a patch to drop autodetect like a =AUTO
option or so (want to let the test do its thing). Imagine Qt will
revert this as well eventually so it won't be there forever.

Bug: https://bugs.gentoo.org/754021
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>