systemd.eclass: make all fallback paths use /usr/lib
Back in 2018, the eclass was changed to ensure that the canonical path
used for results was in /lib, since that works for split-usr systems
running systemd while also supporting merged-usr systems due to portage
following symlinks at install time. It even works for binpkgs.

But here and now in 2025, systemd doesn't support split-usr at all.
There is no point in having unit files install conservatively in
/lib/systemd. Update the path to accommodate the new reality.

This mostly has no effect. On openrc profiles, or for packages that use
systemd_dounit without pulling in pkgconfig, binpkgs will be created
with a different, but still working path, which then triggers iwdevtools
warnings when you reinstall a former binpkg from source, as it thinks
the file has "moved". e.g.

```
 * CMP: =net-misc/radvd-2.19-r7 with net-misc/radvd-2.20/image
 *  FILES:-lib/systemd/system/radvd.service
 *  FILES:+usr/lib/systemd/system/radvd.service
 * ------> FILES(+1,-1)
```

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>