sys-auth/polkit: drop USE=daemon
I really think I just had a braino here and meant to add it for
dbus (for dbus-broker) and somehow ended it up to polkit when bumping
it as I'd had it on my mind, or something.

See be4011d5f8b5d5a01a1bf0677c8863313349963c as well. I can't see
any real use for this other than perhaps avoiding spidermonkey back
before duktape wasn't an option (when packages needed libpolkit but
you really didn't want to run the daemon or could live without it).

The *upstream* intended-use for USE=daemon is for Flatpak where there's
no point in having the daemon as you use the system daemon there but
you need the utilities within a Flatpak.

Closes: https://bugs.gentoo.org/954290
Signed-off-by: Sam James <sam@gentoo.org>