dev-php/pecl-imagick: fixup RDEPEND and update a comment
We're dropping the USE=openmp block on imagemagick, because maybe it
isn't a problem any more. This commit mostly updates a comment in the
ebuild. We also add back a subslot dep on imagemagick that was removed
in a PR. No -r1 because this is being pushed at the same time as the
commit that removes it.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>