media-video/ffmpeg: tighten vulkan dependencies with libplacebo
See bug #954764 comment #2 for explanations.

Technically needs a revbump, but do not want to revbump ffmpeg too much
given still a decently long build that is on many users' systems, and
most of them do not enable libplacebo. At worst, if a user does -vulkan
on dependencies after a successful build, only ffplay or specific
filters will be broken and not the ffmpeg libraries thanks to stubs.

Closes: https://bugs.gentoo.org/954764
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>