dev-libs/zziplib: forcefully set CMAKE_BUILD_TYPE
It's critical that we use RelWithDebInfo and not Release (which a user
may set) because the upstream CMake sets a different library name (!)
with Release.

With that different name (which we don't want based on an environment
variable anyway), the pkg-config file is also wrong, so it's not
usable. But we should be grateful for that, as otherwise, we wouldn't
have noticed.

(Of course, it's very very rare for a user to set this, and in the end,
they've decided to stop doing that after discussing it, but nonetheless...)

Bug: https://github.com/gdraheim/zziplib/issues/117
Thanks-to: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>