dev-qt/qtwebengine: allow using libatomic-stub
Adapted from linked PR given hasn't been updated yet and wanted
to rebuild qtwebengine with the new dependencies on my llvm-musl
chroot now.

[atomic-builtins] is not enough given passes -latomic either way,
and not planning to try to get this fixed in chromium.

Also move to DEPEND-only, doesn't seem to be linked with shared
libatomic for gcc (and stub is static-only).

Skipping revbump given not worth rebuilds just to be able to
depclean gcc and qtwebengine-6.9.1 is not that far off. Besides
it's difficult to depclean given nodejs depends on it for libatomic
at the moment still.

Closes: https://github.com/gentoo/gentoo/pull/41689
Thanks-to: Michal Rostecki <vadorovsky@disroot.org>
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>