dev-libs/boost: apply "keep-going" for tests
Run the whole testsuite and print errors at the end. We generally want
these semantics for testsuites so that e.g. if one failure occurs, we can
look to see if there's another "easier" one to analyse. Plus, people
running the testsuite generally don't mind if it takes a bit longer as long
as it's not unreasonable (people don't broadly want fail-fast for it;
they don't want to know some e.g. linter test failed but not find out later
that some critical core functionality test later didn't fail as it never
got run).

Could maybe trim the test summary failure output to drop the path
but not going to bother for now.

Signed-off-by: Sam James <sam@gentoo.org>