app-emulation/wine-proton: add 10.0.1b
Consider this experimental, may want to stick to wine-proton-9 for
now if do not want games to break.

Changes:
* Tentatively re-add X + wayland IUSE, albeit I cannot particularly
  recommend wayland-only for games at the moment (this did not even
  support GL with wayland last version) and it is unlikey Valve tests
  that configuration
* make_vulkan newly needs `-X video.xml` to avoid downloading it
* pass --without-piper like wine-proton-9999 does, not packaged
* drop vulkan-libm.patch, ceil is no longer used

Beside that, only changes are typical to changes between wine-9
and 10 and are not proton specific. Not much seem to have changed
with proton patches that need ebuild changes albeit it is hard to
say at a glance.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>