app-emulation/wine-*: adjust libglvnd dependencies for wayland
Noticed while working on wine-proton-10 that forgot to setup
the libglvnd dependency in wine-10 for wayland (it newly added
EGL support that was missing in wine-9).

At same time, add a required USE given these are no-op unless
X or wayland is set. fwiw vulkan on the other hand still does
something albeit unsure what it'd be useful for, perhaps headless
filters and such.

Not really worth revbumps, it'd be hard to be missing this and,
even if did, the dlopen nature wouldn't break wine entirely.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>