dev-libs/libcec: remove cubox use flag
These headers may have never existed in the gentoo tree, older versions just
silently compiled causing breakage elsewhere. Now the situation has
improved in 7.0.0 where it will fail to configure without the headers
present atleast. But this doesn't address the fact that these headers are
still elusive.

These headers were most likely available in an android fork of the
kernel at some time and some other related headers are in an out of tree
module beagle-nxp-hdmi which was last updated in 2012 and was never in
the gentoo tree.

Closes: https://bugs.gentoo.org/955124
Closes: https://bugs.gentoo.org/922690
Closes: https://bugs.gentoo.org/818262
Closes: https://bugs.gentoo.org/824978
Signed-off-by: Alfred Wingate <parona@protonmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/41865
Signed-off-by: Sam James <sam@gentoo.org>