x11-misc/xdotool: remove ldconfig patch, skip unused libXi entirely
Xi is being dropped by --as-needed and I didn't see it in DEFAULT_LIBS,
but I missed that it was being added by pkg-config too.

For ldconfig, can use LDCONFIG=: instead.

Also considered using DESTDIR given it exists, but still need to
define lib/man dir and it just becomes longer with ${EPREFIX}.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>