app-shells/bash: add 5.2_p37-r1 with several fixes
Backport two supplemental patches from the devel branch so as to address
some remaining defects affecting the revised read_mbchar() function.
For further information, please refer to pitfall #65 at
https://mywiki.wooledge.org/BashPitfalls, along with the commit messages
of the patches themselves. The read builtin should now behave as well as
it did in 4.4, and as well as it will in the upcoming 5.3-rc2 release.

Whitelist the Contour and Ghostty terminal emulators for colour support.

In the case that colour support is detected, refrain from defining any
aliases for which either an alias or function is found already to exist
by the same name. The motivation for making this change stems from a
discussion at https://forums.gentoo.org/viewtopic-t-1170323.html.

Define COLORTERM=1 in the environment of dircolors(1). The methods by
which dircolors determines whether colour is supported are inferior to
those of Gentoo. Indeed, at the point that dircolors is executed, it
will already have been determined that colour is supported with some
confidence. Defining COLORTERM=1 coerces dircolors into presuming so in
turn, increasing the range of terminal emulators for which Gentoo is
able to define LS_COLORS.

Define _cwd as a local variable in genfun_set_win_title(). Consequently,
_cwd is no longer a reserved variable name from the perspective of the
user.

Whitelist the Contour terminal emulator for XTWINOPS support.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>