app-shells/bash: add 5.2_p37-r2 with optimised colour/prompt intialisation
This revision brings two changes to the 10-gentoo-color.bash drop-in.

Firstly, try to match the value of TERM against the whitelist of
terminals known to support colour, only employing tput(1) as a last
resort. Given that the list is now a well-curated one, it makes sense to
try to forgo the expense of executing an external utility.

Secondly, micro-optimise the method by which readline support is
detected. It is ever so slightly faster to run ( ! shopt -u direxpand )
as opposed to [[ ! $(shopt -p direxpand) ]].

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>