dev-util/bpftool: prevent attribute warning with bpf-toolchain
The bpf-generated kernel headers currently use a clang-specific
pragma to annotate data structures with the preserve_access_index
attribute; since this does not work with gcc, the attribute will
be ignored. Luckily the BPF test suite already has a macro guard
called BPF_NO_PRESERVE_ACCESS_INDEX for this case, so just define
that for the bpftool build as well.

Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>