dev-java/xsdlib: adjust dependencies "Possibly unneeded dependencies found"
There was xerces-2 wrongly registered in the DEPEND line in package.env.
This commit moves dev-java/xerces:2 from CP_DEPEND to DEPEND and adds it
to JAVA_CLASSPATH_EXTRA so that in /usr/share/xsdlib/package.env it gets
registered as BUILD_DEPEND="xerces-2" instead in the DEPEND line.

Also sorts variables and removes useless comments.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>