profiles/arch/x86: force wine-*[mingw] for now
The >=9 still forces it for some ebuilds where it's not
needed, but these are already overshadowed in stable and
due for cleanup so not really worth trying to exclude.

Have not really got to the bottom of this, could potentially
be needed on amd64 under some conditions too. Will probably
know after stabilized on amd64 too.

Closes: https://bugs.gentoo.org/955405
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>