dev-lang/rust: Add `RUST_SYSROOTS` as `USE_EXPAND` variable
Add a new `USE_EXPAND` variable called `RUST_SYSROOTS` which allows to
specify additional Rust sysroots to build. Rust sysroot is a directory
which contains the prebuilt standard libraries (in form of `*.rlib`
files) and utility binaries (e.g. `rust-objcopy`).

This variable is limited to bare metal targets. Sysroot for targets not
falling into that category contain a local copy of compiler runtime,
libc and unwinder, so crossdev is a better fit for them.

In this change, the only `RUST_SYSROOT` option is `WebAssembly`, which
enables build of a sysroot for wasm32-unknown-unknown target.

Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>