eclass/tests: tweak dot-a.sh tests for binutils
I'm not sure if I made a mistake when testing this before or if it changed
when the strip patch did, but the new behaviour is fine too with the
latest patch for Binutils.

I think what happened is that the patch changed to only mangle with -R ...
where ... is for LTO sections, which is fine for us. But doing this in a
separate commit so I can revert it easily if needed.

See https://sourceware.org/bugzilla/show_bug.cgi?id=21479#c13.

Signed-off-by: Sam James <sam@gentoo.org>