dev-libs/newt: cleanup python version selection
Back in 2018, version 0.52.21 got a dedicated configure parameter to
override the `ls /usr/include/python*/Python.h` used by default.

It was anyways pointless to pass it to econf and then individually
override it in every call to make, since the configure script doesn't
even try to look at the existing value. But anyways it's been six years
since it was relevant to use overrides. Get rid of them and use econf.

in fact, passing PYTHON at configure time was totally broken anyway. It
specifically needs EPYTHON, no full paths. This was correct when passing
the values to make.

Bug: https://pagure.io/newt/issue/3
Acked-by: Ben Kohler <bkohler@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>