dev-libs/newt: remove obsolete custom-patched makefile var to set installdir
During the most recent version bump, newt-0.52.21-python-sitedir.patch
was dropped from PATCHES, because upstream stopped installing python
code to --libdir. This represented an upstream fix for the problem the
patch tried to solve.

The downstream patch added a Makefile variable which was unset, and had
to be passed as a `make` parameter. We still passed that variable in
`emake` even though the patch which defined it stopped being applied.
Drop the outdated parameter.

Bug: https://pagure.io/newt/pull-request/26
Acked-by: Ben Kohler <bkohler@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>