dev-libs/newt: make python bindings optional
Supported since 0.52.13, released in 2011. Few packages need these
bindings, and it's simpler to update a system, the fewer random C
libraries attempt to acquire a lock on PYTHON_TARGETS.

Furthermore, newt is required by the command-line tools for
networkmanager, which is quite popular and means it is significantly
more likely that PYTHON_COMPAT *needs* to be kept promptly up to date.

Acked-by: Ben Kohler <bkohler@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>