mail-mta/exim-4.98.2-r1: fix build with C23 by disabling eximon
- bump to EAPI=8
- drop USE=X to disable eximon

eximon is very old, and probably not that much maintained code.  It
isn't enabled by default (upstream) and relies on X, which is dated.  It
doesn't compile with -std=c23, and upstream has the habit of refusing
any fix for compiler warnings/errors, so just disable eximon, as it is
just a client tool, the MTA is unaffected.

Closes: https://bugs.gentoo.org/955368
Signed-off-by: Fabian Groffen <grobian@gentoo.org>