x11-misc/fbpanel: use patch rather than 2to3 + fix w/ gcc-15
Had forgot about this, and it worked by accident when
tested it.

The patch is smaller than was expecting, so tossing it
in filesdir like the rest.

Not that this package has a long life ahead of it, I only
look at it now and then because I know someone on the forums
that is using it (I don't). Next time there is a major issue
I'd say it should be last-rited.

Closes: https://bugs.gentoo.org/955505
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>