flag-o-matic.eclass: allow -ffat-lto-objects in strip-flags
This flag is "safe" for the same reason `-flto` is "safe". Furthermore,
the new dot-a.eclass relies on setting this flag, which `strip-flags` then
deletes again in the event that a package uses both `lto-guarantee-fat` and
`strip-flags`.

End result: static libraries that contain neither LTO nor regular object
code.

Example package that does this: dev-libs/nss

Acked-by: Sam James <sam@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>