Revert "sys-apps/dtc: fix static build"
This reverts commit 0428966db63c13fc0142f0500e5c44895793eb37.

At least, revert it in live. The entire commit was totally bogus. The
build option in question just determined whether to build using
LDFLAGS="-static". Upstream didn't support building without static libs
at all, until

https://git.kernel.org/pub/scm/utils/dtc/dtc.git/commit/?id=dd1b3e532d228ed75d34480ee539f71b76a4fa89

so the correct solution to support USE=static-libs was always, rm the
file after the fact.

But now that it is correctly supported in live we can go back to the
original intention.

Bug: https://bugs.gentoo.org/907940
Bug: https://bugs.gentoo.org/909366
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>