x11-terms/kitty-terminfo: stabilize everywhere
Doing it myself rather than bother arch testers given this does
nothing but run a tic(1) command and is ALLARCHES. The keywords
were similarly added in commit 16f7af268946.

Perhaps a minor annoyance is that will have to put kitty-terminfo
and shell-integration in a separate bug in the future for ALLARCHES
to work and, while not a major issue, being stabilized for ALLARCHES
may print "skipped upgrade" warnings for users until the main kitty
package is stabilized on that arch too due to the pinning which may
be worse if e.g. x86 is done first then amd64/arm64 stalls.

Closes: https://bugs.gentoo.org/955573
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>