dev-build/autoconf-archive: add Lua revert
This seems stalled upstream so go for the lesser evil. There's an
alternative PR upstream as well fixing this in full but reverting
feels better to do downstream than ship a macro variant never
seen before that may end up changed/not merged.

Thought we had a bug for the Lua issues but I don't spot one.

Signed-off-by: Sam James <sam@gentoo.org>