media-libs/libcdr: make sure elibtoolize is applied
... and not just when building live.

Needed to e.g. pass LTO warning flags through to the linker. Also needed
to pass through --as-needed correctly to the linker, which has the
practical effect of pruning a link to some unused lcms plugins.

lcms provides its plugins in the main .pc file by design to "simplify
usage", if configuring to build the plugin at all, and expects
--as-needed to strip those out if you don't call `cmsPlugin`. But that
relies on libtool doing its job...

(We unconditionally package the plugins, so this is just an
optimization.)

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>