app-accessibility/speech-dispatcher: fix automagic systemd support
In this upstream release, user units were added that also link to
libsystemd.so; it is autodetected if both the unit files were installed
AND libsystemd was found via pkg-config.

We need a USE flag to control this. Once we have a USE flag it is
pointless to install the system unit unconditionally (the "small files"
QA rule) since users will anyways want/need socket activation or just build
with global USE=systemd.

Fixes:
```
 * VDB: detected possibly incorrect RDEPEND (app-accessibility/speech-dispatcher-0.12.0)
 *  > sys-apps/systemd
```

Fixes: 1f6eb8673d45d10412a8e10740f6f81587724191
Closes: https://bugs.gentoo.org/955909
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>