x11-libs/wxGTK: bodge ABI issues one more time
See https://bugs.gentoo.org/955902#c8 onwards for details, but in summary,
we have a bunch of seds in the ebuild, and they break the version script upstream
supplies, resulting in made-up symbol versions that are Gentoo-specific (!!!)

```
       sed -i \
               -e "s:\(WX_VERSION=\).*:\1${WXVERSION}:" \
               -e "s:\(WX_RELEASE=\).*:\1${WXRELEASE}:" \
               -e "s:\(WX_SUBVERSION=\).*:\1${WXSUBVERSION}:" \
               -e '/WX_VERSION_TAG=/ s:${WX_RELEASE}:3.0:' \ # <--- here
               configure || die
```

I question how much of the seds are necessary at all, but the mangling of the version
script is a critical ABI issue as it makes us incompatible with other distributions
and upstream binaries.

However, fixing this correctly while also quickly to solve bug #955902
isn't really feasible, so..

Quickly bodge this by patching the affected symbols in
https://github.com/wxWidgets/wxWidgets/commit/f3eb19ce2efa839c0d3e8dc922445cbb49bd9e5d
to have 3.0.x. In for a penny, in for a pound.

Then can try actually fix this once wxGTK is building again and figure out
proper ABI compatibility with our made up symbols as well as the upstream ones
in a followup revision. Filed bug #955936 for that proper fix.

TL;DR: Leave proper fixing for it until later to fix the immediate emergency,
extend the previous hack to newly-added symbols.

Bug: https://bugs.gentoo.org/955936
Closes: https://bugs.gentoo.org/955902
Signed-off-by: Sam James <sam@gentoo.org>