Start elogind before display-manager
I believe this was missed when the switch to unified `display-manager` init script was made. Without it, openrc users can have issues with seats (preventing eg `loginctl reboot` from working) or with pulseaudio (eg having a system wide daemon started as root, preventing bluetooth devices from showing as available outputs once paired/connected).

Of course, there could be a few more things in this PR - do I need to update the copyright notice to 1999-2025, do I need to write a elogind.init-r2 file instead, etc? Whatever the case may be, this change seems to have a positive impact on the aforementioned issues, so I'm putting it out there.

Thanks to `navi` and `negril` on IRC for encouraging me to make the PR :-) and thank you for reading!

Signed-off-by: Cheaterman <the.cheaterman@gmail.com>