dev-libs/optix: make headers-only #955908
Building the SDK part has no purpose outside of building examples. It has no
possible consumers. The source archive needs manual fetch.

Given that there is an upstream repo now we can use as source archive there is
no reason to use the manual fetch SDK archive as alternative source either.

So drop the logic associated with that and make the a headers only package.

Closes: https://bugs.gentoo.org/955908
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>